+++ src/mbgl/util/mat3.cpp
@@ -102,9 +102,9 @@ void scale(mat3& out, const mat3& a, double x, double y) {
 }
 
 void transformMat3f(vec3f& out, const vec3f& a, const mat3& m) {
-    out[0] = m[0] * a[0] + m[3] * a[1] + m[6] * a[2];
-    out[1] = m[1] * a[0] + m[4] * a[1] + m[7] * a[2];
-    out[2] = m[2] * a[0] + m[5] * a[1] + m[8] * a[2];
+    out[0] = static_cast<float>(m[0]) * a[0] + static_cast<float>(m[3]) * a[1] + static_cast<float>(m[6]) * a[2];
+    out[1] = static_cast<float>(m[1]) * a[0] + static_cast<float>(m[4]) * a[1] + static_cast<float>(m[7]) * a[2];
+    out[2] = static_cast<float>(m[2]) * a[0] + static_cast<float>(m[5]) * a[1] + static_cast<float>(m[8]) * a[2];
 }
 
 } // namespace matrix
